v2.1
Started introducing additional compatibility measures, and started
working on the mini-game framework.

Added EnkiBlock and ForgeBlock.

Moved EnkiPlayer to the correct namespace.

Introduced CompatabilityAccess, which allows access to objects
representing various minecrafty things, as the subclasses/in the
formats meant for the API it’s currently being used in. (Forge, Bukkit,
Sponge, etc.).

Added initialisation for CompatabilityAccess to the EnkiLib Forge mod
class.

Added the start of the base PvpGame class.