r1.0.0
Changed versioning system to something less arbitrary and convoluted,
to something more informational and genuinely useful, based on
compatibility.

Removed clone method from the Range and FlatRange interfaces.

Added includesMin, includesMax, and isFlat methods to the Range
interface, allowing ranges to be defined as extending to just-below or
just-above a value, such that it doesn’t include it, but includes
anything below or above it within the range.

Implemented ExclusiveRange and ValueRange. See the v2.1.1 continuance
change logs for more detailed information concerning their
implementation and development.